reorganised and exposed table
