+++ lib/components.test.js
@@ -55,6 +55,7 @@ test("all components exported", () => {
       "StatusDot",
       "StyledLink",
       "SystemThemeProvider",
+      "Table",
       "Tab",
       "TabsContainer",
       "Tag",
