+++ lib/index.js
@@ -51,6 +51,7 @@ export { default as SideNav } from "./components/SideNav";
 export { default as Spacer } from "./components/Spacer";
 export { default as StatusDot } from "./components/StatusDot";
 export { default as StyledLink, styleLink } from "./components/StyledLink";
+export { default as Table } from "./components/Table";
 export { Tab, TabsContainer } from "./components/Tabs";
 export { default as Tag } from "./components/Tag";
 export { default as TextInput } from "./components/TextInput";
